attendance_system/wsgi.py
"""
WSGI config for attendance_system project.

It exposes the WSGI callable as a module-level variable named ``application``.

For more information on this file, see
https://docs.djangoproject.com/en/4.0/howto/deployment/wsgi/
"""

import os
from dotenv import load_dotenv
from pathlib import Path
from django.core.wsgi import get_wsgi_application

BASE_DIR = Path(__file__).resolve().parent.parent
os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'attendance_system.development')
load_dotenv(os.path.join(BASE_DIR, '.env'))

print(f'SECRET {os.getenv("SECRET_KEY")}')
application = get_wsgi_application()
