requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
dj-static==0.0.6
Django==4.0.2
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.11.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.1
openapi-codec==1.3.2
python-dotenv==0.21.0
pytz==2021.3
random-username==1.0.2
rcssmin==1.1.0
requests==2.27.1
rjsmin==1.2.0
simplejson==3.18.0
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
